requirements.txt
asgiref==3.7.2
Django==4.2.5
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-modeltranslation==0.18.11
django-parler==2.3
django-parler-rest==2.2
djangorestframework==3.14.0
Pillow==10.0.0
python-environ==0.4.54
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
